debugging 8

+++ Sorter.c
@@ -62,7 +62,7 @@ int main(int argc, char **argv)
 
 	//METADATA, DO NOT DELETE
 	files = calloc(1, sizeof(struct csv *) * fileCap);
-	printf("Initial TID: %d\nTIDS of all child threads: ", (unsigned)pthread_self());
+	printf("Initial TID: %d\nTIDS of all child threads: ", (unsigned long)pthread_self());
 	printf("\n\n");
 	fflush(stdout);
 	int totalNumThreads = parseDir(directory, outputDirectory, query);
@@ -390,7 +390,7 @@ int parseDir(char *inputDir, char *outputDir, char *sortBy)
 	} 
 	
 	int maxPossibleThreads = 10;
-	unsigned int *listOfThreadIDs = (unsigned int *) malloc(maxPossibleThreads*sizeof(unsigned int));
+	unsigned long *listOfThreadIDs = (unsigned long *) malloc(maxPossibleThreads*sizeof(unsigned long));
 	int numChildThreads = 0;
 	
 	int totalNumThreads = 1;
@@ -410,7 +410,7 @@ int parseDir(char *inputDir, char *outputDir, char *sortBy)
 			sortFileParameters->fileName = pDirent->d_name;
 			sortFileParameters->sortBy = sortBy;
 			pthread_create(&tid, NULL, threadExecuteSortFile, (void *)sortFileParameters);
-			printf("tid=%d\n", (unsigned)tid);
+			printf("tid=%d\n", (unsigned long)tid);
 			if (numChildThreads<maxPossibleThreads)
 			{
 				printf("Flag2\n");
@@ -421,7 +421,7 @@ int parseDir(char *inputDir, char *outputDir, char *sortBy)
 			{
 				printf("Flag3\n");
 				maxPossibleThreads = maxPossibleThreads*2;
-				unsigned int *tempPtr= (unsigned int *)realloc(listOfThreadIDs, maxPossibleThreads);
+				unsigned long *tempPtr= (unsigned long *)realloc(listOfThreadIDs, maxPossibleThreads);
 				listOfThreadIDs = tempPtr;
 				listOfThreadIDs[numChildThreads] = tid;
 				numChildThreads++;
@@ -461,7 +461,7 @@ int parseDir(char *inputDir, char *outputDir, char *sortBy)
 			{
 				printf("Flag5\n");
 				maxPossibleThreads = maxPossibleThreads*2;
-				unsigned int *tempPtr= (unsigned int *)realloc(listOfThreadIDs, maxPossibleThreads);
+				unsigned long *tempPtr= (unsigned long *)realloc(listOfThreadIDs, maxPossibleThreads);
 				listOfThreadIDs = tempPtr;
 				listOfThreadIDs[numChildThreads] = tid;
 				numChildThreads++;
@@ -481,7 +481,7 @@ int parseDir(char *inputDir, char *outputDir, char *sortBy)
 	printf("Total of numChildThreads=%d\n", numChildThreads);
 	for (i=0;i<numChildThreads;i++) 
 	{
-		printf("Join %d number=%d\n", i, (unsigned)listOfThreadIDs[i]);
+		printf("Join %d number=%d\n", i, (unsigned long)listOfThreadIDs[i]);
 		pthread_join(listOfThreadIDs[i], (void *)&status);  //blocks execution until thread is joined
 		totalNumThreads += status;
 	}
